1.4.4

+++ core/domain/src/main/kotlin/org/michaelbel/movies/domain/usecase/MovieDetailsCase.kt
@@ -1,14 +1,14 @@
 package org.michaelbel.movies.domain.usecase
 
 import javax.inject.Inject
-import org.michaelbel.movies.domain.interactor.MovieInteractor
+import org.michaelbel.movies.domain.interactor.movie.MovieInteractor
 import org.michaelbel.movies.entities.handle
 import org.michaelbel.movies.entities.lce.ScreenState
 
 class MovieDetailsCase @Inject constructor(
     private val movieInteractor: MovieInteractor
 ) {
-    suspend operator fun invoke(movieId: Long): ScreenState {
+    suspend operator fun invoke(movieId: Int): ScreenState {
         movieInteractor.movieDetails(movieId).handle(
             success = { movieDetailsData ->
                 return ScreenState.Content(movieDetailsData)
